Split parsing into its own class
